Add files via upload
Data simulation updated to Version 1.01 (skipping 1.00 which was the final data sim with the 10/5 update after first review of the walkthrough document. The jump to version 1.00 notes the first completed data simulation that was used in finalizing a walkthrough notebook and this use case's first presented/presentable iteration). Version 1.01 includes only updates to the code's preface, explicitly acknowledging the proposed mortality status varible has not been created.